Extracted infinite canvas from the main content view and put into its own BoardCanvasView, added toolbar overlay on top also
